+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Domain/ChargeLinkTransmissionRequest/IChargeLinkTransmissionRequestFactory.cs
@@ -13,20 +13,16 @@
 // limitations under the License.
 
 using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 using GreenEnergyHub.Charges.Domain.ChargeLinkCommandAcceptedEvents;
 using GreenEnergyHub.Charges.Domain.MarketParticipants;
 
-namespace GreenEnergyHub.Charges.Domain.ChargeLinkHistory
+namespace GreenEnergyHub.Charges.Domain.ChargeLinkTransmissionRequest
 {
-    public interface IChargeLinkTransmissionRequestRepository
+    public interface IChargeLinkTransmissionRequestFactory
     {
-        Task StoreAsync(
+        ChargeLinkTransmissionRequest MapChargeLinkCommandAcceptedEvent(
             ChargeLinkCommandAcceptedEvent chargeLinkCommandAcceptedEvent,
             MarketParticipant marketParticipant,
             Guid messageHubId);
-
-        Task<List<ChargeLinkTransmissionRequest>> GetChargeLinkTransmissionRequestsAsync(IEnumerable<Guid> postOfficeIds);
     }
 }
