+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Infrastructure/Context/ChargesDatabaseContext.cs
@@ -14,8 +14,8 @@
 
 using System;
 using System.Threading.Tasks;
-using GreenEnergyHub.Charges.Domain.ChargeLinkHistory;
 using GreenEnergyHub.Charges.Domain.ChargeLinks;
+using GreenEnergyHub.Charges.Domain.ChargeLinkTransmissionRequest;
 using GreenEnergyHub.Charges.Domain.MeteringPoints;
 using GreenEnergyHub.Charges.Infrastructure.Context.EntityConfigurations;
 using GreenEnergyHub.Charges.Infrastructure.Context.Model;
